Basic, functioning grammar of graphics in Go
There are huge holes in this implementation, and many of the things
that are here aren't right, but the key pieces are in place and they
fit together well enough to plot a simple curve to an SVG. There is a
notion of a plot with layers, marks, transforms, and scales. It has a
fairly complex implementation of nominal, ordinal, and cardinal
variables.