gg: make AutoVar sensible, eliminate NewVar1
Currently AutoVar returns a single-element Var for constant data. This
makes everything else harder because even if that type is cardinal,
orderable, or comparable, it's just a Var and attempting to cast it to
a more specific type will fail.

Previously, I represented these using a special VarConst that didn't
fit in the Var hierarchy because there was only one element and thus
all of the operations were degenerative. Having to deal with this in
everything was annoying, so I switched to cycling vars and introduced
Var1s to replace VarConst. However, now that there is a single Var
hierarchy and any Var subtype can represent "constant" data by cycling
it, it makes much more sense to use that hierarchy.

Hence, this commit eliminates Var1s and makes AutoVar return the most
specific Var subtype for constants just as it does for sequences.