gg: make TransformSort consistent
This fixes various inconsistencies in TransformSort. First, it makes
it cycle or slice all bindings to the length of the binding being
sorted by so the permutation never fails and the behavior matches that
of layers. It also makes it require at least an ordinal variable,
regardless of its length, which agrees with the new behavior of
AutoVar.