gg: grouping support
This changes a binding from being a single Var with a single Scaler to
being a set of (Var, Scaler) pairs identified by a group ID. All
bindings must have a consistent set of groups, and binding a Var (or
an AutoVar) broadcasts that Var to all groups.

We modify all transforms and layers to map across all groups. Several
operations that used to take a binding now take a binding group.