table: make the empty Table also the empty Grouped
Currently the empty Table still has the root group, which makes it
different from the empty Grouped, which has no groups. This means we
need a separate notion of an empty Grouped.

Simplify things by defining the empty Table to have no groups, which
makes it also the empty Grouped.