gg: switch to using table package
This replaces the implicit data table implied by the current set of
Plot bindings with a first-class, explicit data table represented as a
table.Grouping.

This simplifies a large number of things in the process. There's no
longer a notion of cycling; the table forces all columns to have the
same length and there are separate constant bindings (though these
aren't in the public API yet). Many of the harrier operations have
been offloaded to the table package (where they're also much simpler)
and now use only public APIs. Vars, possibly one of the most complex
parts of the old implementation and interface, are simply gone;
replaced with better use of the Go type system.