table: fix GroupID of first Table added to a group
Currently if a Table is added to an empty Table to produce a Grouped,
we lose the GroupID of the added Table because we short-circuit and
return the added Table, which is always in the root group. This
optimization is only safe if the Table being added is specifically
given the RootGroupID. Fix the optimization accordingly.