gg: restructure layering and marking
This more clearly separates the two basic roles of a layer: training
scales and marking. In particular, the layer function itself is now
trivial; it merely uses the bindings of interest and registers a mark.
The mark function is also simpler now, mostly because the common path
drawing code has been abstracted out (not because of any fundamental
interface change).