gg: make scales global
Currently scales are scoped so the scale bindings and can be saved and
restored, but this doesn't make much sense. You want the scales to
stay the same as the data and bindings change. This also complicates
questions of where we bind a default scale when we create one:
currently we bind it in the root environment, but this is inconsistent
with user-created scale bindings.

Fix this confusion by making scales global to the Plot.