gg: remove data bindings, layers as structs
This eliminates the concept of data bindings from Plot. Instead,
layers are structs and are configured simply by setting fields in
their structs to column names (or anything else). This also makes X
and Y default to the first and second columns, which meshes nicely
with the zero value of these structs.

This significantly reduces the API surface of Plot. It eliminates the
somewhat confusing intermediate level of indirection between columns
and aesthetics, replacing it with a perfectly natural language
mechanism. Property names are statically checked. And there's no
question of what happens to bindings when a new table is set because
they don't persist in the Plot.

On the other hand, it's more awkward to work with than I had expected,
mostly because compound literals can't directly initialize fields of
embedded structs, so it's difficult to wrap an existing layer in a new
type to create a variant of that layer. It's also potentially less
flexible, since I can't combine properties from multiple sources, such
as coordinate space-defined properties or renderer-defined properties.