table: make group labels any value
Currently group labels are strings, but this is restrictive and makes
them useless for carrying semantic information.

Change group labels to be interface{}. This lets us efficient store
information for useful formatting later; for example, GroupBy now
stores the grouped value so it can be formatted if the table is
printed. It also lets callers that create their own groups store
semantic information in the group hierarchy.