generic: restrict most operations to slices
Currently, several operations accept any of a slice, array, or pointer
to an array. Besides complicating the implementation and the API, this
makes the API inconsistent: for slices, operations can return new
slices of the exact same type, but for arrays and pointers to arrays
this isn't possible if the type is wrapped.

So, for out-of-place operations (which is everything but sort),
simplify the API by only accepting slices.