gg: layout facet labels properly
Currently facet labels are shown on each and every subplot, rather
than on logical rows and columns. This is obviously not what's
desired, but it's tricky to fix in the current design.

Hence, switch to a new design where faceting operations have to record
the global (x, y) coordinate of each subplot and group subplots into
vertical and horizontal "bands". Labels are associated with bands,
rather than individual subplots. This turns out to be fairly easy for
faceting operations to do, while reconstructing this band structure
from per-subplot labels after the fact is a pain. From this,
constructing the plot layout with properly placed facet labels is
relatively clean.