gg: implement facet scale splitting and rework default scales
This changes the way that default scales work so that facet scale
splitting can be implemented reasonably. Currently, default scales are
created when data is first used without a scale. However, if a
faceting operation happens before this, there's no scale for it to
split. Hence, this introduces a new pseudo-Scaler representing a
"default scale" that picks its underlying scalar when its domain is
first expanded. We put this default scale at the root of all scale
trees, so there's always a scale.

Using this, we can implement facet scale splitting by mapping each
distinct scale in each band to a new scale.