gg: add 5% margin within plot area
This shrinks the X and Y rangers by a small amount so that the
extremes of the data (and their tick labels) do not appear right at
the edge of the plot area.

This also fixes the problem where the X and Y axes sometimes don't
quite meet up. Previously these were drawn as two separate lines, but
they basically can't be with plot margins, so now they're drawn as one
line.