gg: don't reuse scaledData for different scalers
Currently the cached scaledData is indexed only by its data source.
This is wrong because the same input data may be used with different
Scalers and this may return a scaledData bound to a different Scaler.

Add the aesthetic to the cache key. This still isn't quite right,
since it should be keyed off the actual scale tree for that aesthetic,
but it's much closer.