gg: initial support for discrete rangers
This gives Ranger two sub-interfaces: ContinuousRanger and
DiscreteRanger. This unification simplifies the Scaler interface,
since it no longer needs two competing methods for setting rangers,
and it lets the discrete ranger interface be more expressive and
dynamic than a simple slice of values.

It updates the linear scale is updated to support discrete rangers and
adds a simple discrete color ranger.