gg: fix multiple plotting of the same mark
Currently the renderer attaches each mark to each subplot it appears
in. However, the mark carries the full set of groups it applies to, so
if a mark appears in multiple groups in the same subplot, it will be
attached to that subplot multiple times. Then, we when render each
subplot, we render each copy of each mark in all of the groups it
appears in, causing each mark to be rendered n^2 times.

Fix this by splitting up the group lists as we associate the marks
with each subplot. This way each mark appears only once in each
subplot, and we can directly render the groups that belong in that
subplot without recomputing them.