gg: allow setting a Ranger on a defaultScale without creating a scale
Currently, setting a Ranger on a defaultScale without an underlying
scale will assume that the underlying scale should be a linear scale.
This is annoying since it means a package user can't in general set
the Ranger of a scale before training it on data, in case the data
implies a different type of scale.

Instead of forcing an underlying scale in this situation, queue up the
Ranger in the defaultScale and pass it on to the underlying scale once
it is created.