ggstat: make ECDF a struct and awesomer
ECDF was the first stat I wrote, so it used a function API instead of
a struct. This turned out to be a bad idea for other stats, so they
use structs with lots of fields to tweak the behavior of that stat.

Catch ECDF up by converting it to a struct and improving it. It now
supports the same widening controls that LeastSquares and LOESS have.
It also supports outputting an additional column with cumulative count
(or weight), and the caller can pass a label to help construct better
column names.