gg: control for distance between tick labels
Currently we find ticks by setting a minimum pixel distance between
the ticks. However, if the labels are long or the font size is large,
the tick labels can collide.

Instead, use a layout-aware predicate function to select ticks with a
minimum distance between the *labels*. This also balanced X and Y
ticks better because the distance between X labels is a function of
the width of the label, while the distance between the Y labels is
only a function of the text height.