table: make Add replace columns in place
Currently Add and AddConst always put the new column at the end of the
table, removing it first if necessary. This isn't actually very
convenient when you care about order, so change Add and AddConst to
replace existing columns in place.

It's easy enough to achieve the old behavior if desired by first
removing the column (whereas emulating the other way around is quite
difficult).