table: switch to builder/viewer API
Currently table exposes only immutable types. "Adding" a column to a
Table returns a new Table. This turns out to be quite inefficient, and
some internal operations like grouping are now using an internal
mutable API to work around this inefficiency.

Switch to a different model in which Tables and Groupings are not only
immutable, but offer no means of expansion. Instead, two new APIs,
Builder and GroupingBuilder allow callers to mutably construct new
Tables and Groupings (but not observe them until they're done). This
turns out to be just as convenient as the functional API in almost all
cases and significantly higher performance.