gg: introduce Unscaled
This makes it possible to wrap values that should be ranged, but not
scaled.

This may not be the right approach. For example, it's mainly useful
for things like opacity, but the results are confusing because the
opacity ranger does *not* go from [0, 1], so you're not actually
specifying the exact opacity.