gg: make Ticks work for non-float64 Scalers
Currently Scaler.Ticks forces the predicate to take tick locations as
float64s. This works for linearScale, but doesn't generalize to
anything else. Related to this, linearScale.Ticks always returns
[]float64s even if its domain is some other numeric type. This happens
to work because its Map method always accepts float64s, but this
clearly won't generalize.

Fix this by doing everything in the domain type. Make the predicate
take ticks in the domain type and return ticks in the domain type.
While we're here, document Ticks in detail to force ourselves to flesh
out the specification.

This makes use of a change to the go-moremath/scale API and hence
depends on at least commit 620232e (2016-08-09).