ggstat: introduce Function, unify Density, LeastSquares, and LOESS
Currently Density, LeastSquares, and LOESS have slightly different
definitions of common fields and slightly different implementations.
For example, Density.Widen is in terms of the KDE bandwidth, while the
other two are in terms of the data span. Similarly, Density keeps the
sample X points as the same type as the input X column, while the
other two always produce []float64 columns.

Fix this by introducing a Function stat that computes univariate
continuous functions over the data and rewriting Density,
LeastSquares, and LOESS to use Function. This unifies the definitions
of N, Widen, and SplitGroups, and makes all three retain the type of
the X column. This also eliminates evalPoints, which was not flexible
enough to express the requirements of all three stats.