ggstat: generalize domain computation to an interface
Currently, each stat that needs a domain duplicates the Widen and
SplitGroups fields and either reimplements the domain computation, or
passes it off to Function. This duplicates code, documentation, and
concepts, and limits the flexibility of domain computation.

Fix this by generalizing domain computation into a FunctionDomainer
interface and replacing all Widen/SplitGroups fields with a
FunctionDomainer. We provide two implementations of FunctionDomainer:
one that implements the current Widen/SplitGroups logic, and another
that implements a simple fixed domain.